Refactored browsers methods, combine them into a individual 'HandleBrowsers' trait
