+++ src/Repositories/ModuleRepository.php
@@ -5,6 +5,7 @@ namespace A17\Twill\Repositories;
 use A17\Twill\Models\Behaviors\HasMedias;
 use A17\Twill\Models\Behaviors\Sortable;
 use A17\Twill\Repositories\Behaviors\HandleDates;
+use A17\Twill\Repositories\Behaviors\HandleBrowsers;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\App;
@@ -16,7 +17,7 @@ use PDO;
 
 abstract class ModuleRepository
 {
-    use HandleDates;
+    use HandleDates, HandleBrowsers;
 
     /**
      * @var \A17\Twill\Models\Model
@@ -527,12 +528,6 @@ abstract class ModuleRepository
      */
     public function afterSave($object, $fields)
     {
-        if (property_exists($this->model, 'browsers')) {
-            foreach ($this->model->browsers as $moduleName) {
-                $this->updateBrowser($object, $fields, $moduleName);
-            }
-        }
-
         foreach (class_uses_recursive(get_called_class()) as $trait) {
             if (method_exists(get_called_class(), $method = 'afterSave' . class_basename($trait))) {
                 $this->$method($object, $fields);
@@ -589,12 +584,6 @@ abstract class ModuleRepository
     public function getFormFields($object)
     {
         $fields = $object->attributesToArray();
-
-        if (property_exists($this->model, 'browsers')) {
-            foreach ($this->model->browsers as $relationship) {
-                $fields['browsers'][$relationship] = $this->getFormFieldsForBrowser($object, $relationship);
-            }
-        }
 
         foreach (class_uses_recursive(get_called_class()) as $trait) {
             if (method_exists(get_called_class(), $method = 'getFormFields' . class_basename($trait))) {
@@ -668,50 +657,6 @@ abstract class ModuleRepository
 
     /**
      * @param \A17\Twill\Models\Model $object
-     * @param string $relation
-     * @param string|null $routePrefix
-     * @param string $titleKey
-     * @param string|null $moduleName
-     * @return array
-     */
-    public function getFormFieldsForBrowser($object, $relation, $routePrefix = null, $titleKey = 'title', $moduleName = null)
-    {
-        return $object->$relation->map(function ($relatedElement) use ($titleKey, $routePrefix, $relation, $moduleName) {
-            return [
-                'id' => $relatedElement->id,
-                'name' => $relatedElement->titleInBrowser ?? $relatedElement->$titleKey,
-                'edit' => moduleRoute($moduleName ?? $relation, $routePrefix ?? '', 'edit', $relatedElement->id),
-                'endpointType' => $relatedElement->getMorphClass(),
-            ] + (classHasTrait($relatedElement, HasMedias::class) ? [
-                'thumbnail' => $relatedElement->defaultCmsImage(['w' => 100, 'h' => 100]),
-            ] : []);
-        })->toArray();
-    }
-
-    /**
-     * @param \A17\Twill\Models\Model $object
-     * @param string $relation
-     * @return array
-     */
-    public function getFormFieldsForRelatedBrowser($object, $relation)
-    {
-        return $object->getRelated($relation)->map(function ($relatedElement) {
-            return ($relatedElement != null) ? [
-                'id' => $relatedElement->id,
-                'name' => $relatedElement->titleInBrowser ?? $relatedElement->title,
-                'endpointType' => $relatedElement->getMorphClass(),
-            ] + (($relatedElement->adminEditUrl ?? null) ? [] : [
-                'edit' => $relatedElement->adminEditUrl,
-            ]) + (classHasTrait($relatedElement, HasMedias::class) ? [
-                'thumbnail' => $relatedElement->defaultCmsImage(['w' => 100, 'h' => 100]),
-            ] : []) : [];
-        })->reject(function ($item) {
-            return empty($item);
-        })->values()->toArray();
-    }
-
-    /**
-     * @param \A17\Twill\Models\Model $object
      * @param array $fields
      * @param string $relationship
      * @param string $formField
@@ -733,49 +678,6 @@ abstract class ModuleRepository
         } else {
             $object->$relationship()->delete();
         }
-    }
-
-    /**
-     * @param \A17\Twill\Models\Model $object
-     * @param array $fields
-     * @param string $relationship
-     * @param string $positionAttribute
-     * @return void
-     */
-    public function updateOrderedBelongsTomany($object, $fields, $relationship, $positionAttribute = 'position')
-    {
-        $fieldsHasElements = isset($fields['browsers'][$relationship]) && !empty($fields['browsers'][$relationship]);
-        $relatedElements = $fieldsHasElements ? $fields['browsers'][$relationship] : [];
-        $relatedElementsWithPosition = [];
-        $position = 1;
-        foreach ($relatedElements as $relatedElement) {
-            $relatedElementsWithPosition[$relatedElement['id']] = [$positionAttribute => $position++];
-        }
-
-        $object->$relationship()->sync($relatedElementsWithPosition);
-    }
-
-    /**
-     * @param \A17\Twill\Models\Model $object
-     * @param array $fields
-     * @param string $relationship
-     * @param string $positionAttribute
-     * @return void
-     */
-    public function updateBrowser($object, $fields, $relationship, $positionAttribute = 'position')
-    {
-        $this->updateOrderedBelongsTomany($object, $fields, $relationship, $positionAttribute);
-    }
-
-    /**
-     * @param mixed $object
-     * @param array $fields
-     * @param string $browserName
-     * @return void
-     */
-    public function updateRelatedBrowser($object, $fields, $browserName)
-    {
-        $object->sync($fields['browsers'][$browserName] ?? [], $browserName);
     }
 
     /**
